# Switch to sync.Pool.